[Refactor] 테스트케이스 구조 변경
Authority, Address, MemberAddress, Tier 테스트케이스 구조변경
